bfloat16!

+++ start_training.py
@@ -98,7 +98,7 @@ class LossMonitor:
 
 if __name__ == "__main__":
     print(f"🚀 Initializing Dynamic Latent Reasoner (Dim={LATENT_DIM})...")
-    model = UniversalReasoner(LATENT_DIM, nnx.Rngs(42), dtype=jnp.bfloat16)
+    model = UniversalReasoner(LATENT_DIM, nnx.Rngs(42), dtype=jnp.float32)
     optimizer = nnx.Optimizer(model, optimizer_chain) 
     
     data_gen = TextDataGenerator(MAX_SEQ_LEN)
